# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trusted IMEI Check API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummary"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummary"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /orders/{orderId}/checkout:
    post:
      operationId: createCheckoutSession
      tags: [orders, stripe]
      summary: Create Stripe Checkout session for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Order not eligible for checkout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /orders/{orderId}/results:
    get:
      operationId: getOrderResults
      tags: [orders]
      summary: Get order results (after payment)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Results available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResults"
        "402":
          description: Payment required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    CreateOrderBody:
      type: object
      properties:
        email:
          type: string
          format: email
        identifier:
          type: string
        identifierType:
          type: string
          enum: [imei, serial]
          default: imei
      required: [email, identifier]
    OrderSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        identifierType:
          type: string
          enum: [imei, serial]
        identifierMasked:
          type: string
        amount:
          type: integer
        currency:
          type: string
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        checkStatus:
          type: string
          enum: [pending, in_progress, completed, failed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, identifierType, identifierMasked, amount, currency, paymentStatus, checkStatus, createdAt, updatedAt]
    CheckoutSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
      required: [sessionId, url]
...
    OrderResults:
      type: object
      properties:
        orderId:
          type: string
        checkedAt:
          type: string
          format: date-time
          nullable: true
        identifierType:
          type: string
          enum: [imei, serial]
        identifierMasked:
          type: string
        providerName:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        manufacturer:
          type: string
          nullable: true
        blacklistStatus:
          type: string
          enum: [clear, blacklisted, unavailable, error]
          nullable: true
        activationLockStatus:
          type: string
          enum: [on, off, unavailable, error]
          nullable: true
        findMyStatus:
          type: string
          enum: [on, off, unavailable, error]
          nullable: true
        providerCoverageNotes:
          type: string
          nullable: true
        checkStatus:
          type: string
          enum: [pending, in_progress, completed, failed]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
      required: [orderId, identifierType, identifierMasked, checkStatus, paymentStatus]
...