# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trusted IMEI Check API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyIdentifierType } from "./createOrderBodyIdentifierType";

export interface CreateOrderBody {
  email: string;
  identifier: string;
  identifierType?: CreateOrderBodyIdentifierType;
}