# lib/api-zod/src/generated/types/orderResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trusted IMEI Check API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderResultsActivationLockStatus } from "./orderResultsActivationLockStatus";
import type { OrderResultsBlacklistStatus } from "./orderResultsBlacklistStatus";
import type { OrderResultsCheckStatus } from "./orderResultsCheckStatus";
import type { OrderResultsFindMyStatus } from "./orderResultsFindMyStatus";
import type { OrderResultsIdentifierType } from "./orderResultsIdentifierType";
import type { OrderResultsPaymentStatus } from "./orderResultsPaymentStatus";

export interface OrderResults {
  orderId: string;
  checkedAt?: Date | null;
  identifierType: OrderResultsIdentifierType;
  identifierMasked: string;
  providerName?: string | null;
  brand?: string | null;
  model?: string | null;
  manufacturer?: string | null;
  blacklistStatus?: OrderResultsBlacklistStatus;
  activationLockStatus?: OrderResultsActivationLockStatus;
  findMyStatus?: OrderResultsFindMyStatus;
  providerCoverageNotes?: string | null;
  checkStatus: OrderResultsCheckStatus;
  paymentStatus: OrderResultsPaymentStatus;
}